test blendshape map

+++ src/20_SkeletalAnimation/GLWidget.h
@@ -43,7 +43,7 @@ protected:
 
     void glSetting();
 
-    int precision = 64;
+    int precision = 128;
     int iScaleFactor = 0;
     QVector<QOpenGLTexture *> blendShapeTextures;
     void createBlendShapeTex();
